refactor(resolve): split tsoa type logics into transformers
