+++ packages/cli/src/metadataGeneration/transformer/dateTransformer.ts
@@ -1,0 +1,28 @@
+import * as ts from 'typescript';
+import { Tsoa } from '@tsoa/runtime';
+
+import { Transformer } from './transformer';
+import { getJSDocTagNames } from '../../utils/jsDocUtils';
+
+export class DateTransformer extends Transformer {
+  public transform(parentNode?: ts.Node): Tsoa.DateType | Tsoa.DateTimeType {
+    if (!parentNode) {
+      return { dataType: 'datetime' };
+    }
+    const tags = getJSDocTagNames(parentNode).filter(name => {
+      return ['isDate', 'isDateTime'].some(m => m === name);
+    });
+
+    if (tags.length === 0) {
+      return { dataType: 'datetime' };
+    }
+    switch (tags[0]) {
+      case 'isDate':
+        return { dataType: 'date' };
+      case 'isDateTime':
+        return { dataType: 'datetime' };
+      default:
+        return { dataType: 'datetime' };
+    }
+  }
+}
